Refactor: extracted data fetching to useFetchProducts hook
